requirements.txt
openai>=1.30.0
gspread>=5.7.2
oauth2client>=4.1.3
pydrive2>=1.15.0
pdfplumber>=0.10.3
pandas>=2.0.3
